% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ambrosia}
\alias{ambrosia}
\alias{angustifolia}
\alias{chenopodium}
\title{TODO}
\format{
TODO
\describe{
  TODO
}

An object of class \code{matrix} (inherits from \code{array}) with 62 rows and 3 columns.

An object of class \code{matrix} (inherits from \code{array}) with 100 rows and 3 columns.
}
\source{
TODO
}
\usage{
ambrosia

angustifolia

chenopodium
}
\description{
TODO.
}
\details{
TODO
}
\examples{
# TODO
}
\references{
TODO
}
\keyword{datasets}
