% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_test.R
\docType{data}
\name{avail_tests}
\alias{avail_tests}
\alias{avail_cir_tests}
\alias{avail_sph_tests}
\title{Available circular and (hyper)spherical uniformity tests}
\format{
An object of class \code{character} of length 32.

An object of class \code{character} of length 17.
}
\usage{
avail_cir_tests

avail_sph_tests
}
\value{
A character vector whose elements are valid inputs for the
\code{type} argument in \code{\link{unif_test}}, \code{\link{unif_stat}},
\code{\link{unif_stat_distr}}, and \code{\link{unif_stat_MC}}.
\code{avail_cir_tests} provides the available circular tests and
\code{avail_sph_tests} the (hyper)spherical tests.
}
\description{
Listing of the tests implemented in the \code{\link{sphunif}}
package.
}
\examples{
# Circular tests
avail_cir_tests

# Spherical tests
avail_sph_tests
}
\keyword{datasets}
