% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chisq}
\alias{chisq}
\alias{d_chisq}
\alias{p_chisq}
\title{Density and distribution of a chi squared}
\usage{
d_chisq(x, df)

p_chisq(x, df)
}
\arguments{
\item{x}{a vector of size \code{nx} or a matrix of size \code{c(nx, 1)}.}

\item{df}{degrees of freedom.}
}
\value{
A matrix of size \code{c(nx, 1)} with the evaluation of the
density or distribution function at \code{x}.
}
\description{
Computation of the density and distribution functions of a chi
squared.
}
\examples{
curve(sphunif:::d_chisq(x, df = 2), from = 0, to = 5, n = 200)
curve(sphunif:::p_chisq(x, df = 2), add = TRUE, col = 2)
}
\keyword{internal}
