% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{log_besselI_scaled_asymp}
\alias{log_besselI_scaled_asymp}
\title{Evaluation of \eqn{\log(e^{-x} \mathcal{I}_{\nu}(x))} with asymptotics}
\usage{
log_besselI_scaled_asymp(nu, x)
}
\arguments{
\item{nu}{a scalar with the order of the Bessel function.}

\item{x}{vector with evaluation points for the Bessel function.}
}
\value{
A vector of size \code{length(x)} with the evaluated function.
}
\description{
Computes the logarithm of the scaled modified Bessel function
of the first kind with automatic handling of asymptotic cases.
}
\details{
For \code{x} larger than \code{1e4}, an asymptotic approximation
with \code{\link[Bessel]{besselIasym}} is used. For \code{nu} larger than
\code{100}, an asymptotic approximation with
\code{\link[Bessel]{besselI.nuAsym}} is used.
}
\examples{
curve(log_besselI_scaled_asymp(nu = 0.5, x = x), from = 0.01, to = 10)
}
