% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_each_col}
\alias{sort_each_col}
\alias{sort_index_each_col}
\title{Sort the columns of a matrix}
\usage{
sort_each_col(A)

sort_index_each_col(A)
}
\arguments{
\item{A}{a matrix of arbitrary dimensions.}
}
\value{
A matrix with the same dimensions as \code{A} such that each
column is sorted increasingly (for \code{sort_each_col}) or contains the
sorting indexes (for \code{sort_index_each_col}).
\code{sort_index_each_col}.
}
\description{
Convenience functions to sort the columns of a matrix in an
increasing way.
}
\examples{
# Simple example
set.seed(98720222)
A <- matrix(sample(20), nrow = 4, ncol = 5)

# Apply sort by column
sphunif:::sort_each_col(A)
apply(A, 2, sort)

# Apply order by column
sphunif:::sort_index_each_col(A)
apply(A, 2, order)

# Apply ranks by column
sphunif:::sort_index_each_col(sphunif:::sort_index_each_col(A))
apply(A, 2, rank)
}
\keyword{internal}
