% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif-alts.R
\name{r_alt}
\alias{r_alt}
\title{Sample non-uniformly distributed spherical data}
\usage{
r_alt(n, p, M = 1, scenario = "vMF", kappa = 1, nu = 0.5,
  F_inv = NULL, K = 1000)
}
\arguments{
\item{n}{sample size.}

\item{p}{integer giving the dimension of the ambient space \eqn{R^p} that
contains \eqn{S^{p-1}}.}

\item{M}{number of samples of size \code{n}. Defaults to \code{1}.}

\item{scenario}{simulation scenario, must be \code{"vMF"}, \code{"MvMF"},
\code{"ACG"}, \code{"SC"} or \code{"W"}. See details below.}

\item{kappa}{non-negative parameter measuring the strength of the deviation
with respect to uniformity.}

\item{nu}{projection along \eqn{{\bf e}_p}{e_p} controlling the modal
strip of the small circle distribution. Must be in (-1, 1). Defaults to
\code{0.5}.}

\item{F_inv}{quantile function returned by \code{\link{F_inv_from_f}}. Used
for \code{"SC"} and \code{"W"}. Computed by internally if \code{NULL}
(default).}

\item{K}{number of equispaced points on \eqn{[-1, 1]} used for evaluating
\eqn{F^{-1}} and then interpolating. Defaults to \code{1e3}.}
}
\description{
Simple simulation of prespecified non-uniform spherical
distributions: von Mises--Fisher (vMF), Mixture of vMF (MvMF),
Angular Central Gaussian (ACG), Small Circle (SC) or Watson (W).
}
\details{
The parameter \code{kappa} is used as \eqn{\kappa} in the following
distributions:
\itemize{
  \item \code{"vMF"}: von Mises--Fisher distribution with concentration
  \eqn{\kappa} and directional mean \eqn{{\bf e}_p = (0, 0, \ldots, 1)}{
  e_p = (0, 0, \ldots, 1)}.
  \item \code{"MvMF"}: equally-weighted mixture of \eqn{p} von Mises--Fisher
  distributions with common concentration \eqn{\kappa} and directional means
  \eqn{{\bf e}_1, \ldots, {\bf e}_p}{e_1, \ldots, e_p}.
  \item \code{"ACG"}: Angular Central Gaussian distribution with diagonal
  shape matrix with diagonal given by
  \deqn{(1, \ldots, 1, 1 + \kappa) / (p + \kappa).}
  \item \code{"SC"}: Small Circle distribution with axis mean
  \eqn{{\bf e}_p = (0, 0, \ldots, 1)}{e_p = (0, 0, \ldots, 1)} and
  concentration \eqn{\kappa} about the projection along the mean, \eqn{\nu}.
  \item \code{"W"}: Watson distribution with axis mean
  \eqn{{\bf e}_p = (0, 0, \ldots, 1)}{e_p = (0, 0, \ldots, 1)} and
  concentration \eqn{\kappa}. The Watson distribution is a particular case
  of the Bingham distribution.
}

Much faster sampling for \code{"SC"} and \code{"W"} is achieved providing
\code{F_inv}, see examples.
}
