% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_each_col}
\alias{sort_each_col}
\alias{sort_index_each_col}
\title{Sort the columns of a matrix}
\usage{
sort_each_col(A)

sort_index_each_col(A)
}
\arguments{
\item{A}{a matrix of arbitrary dimensions.}
}
\value{
A matrix with the same dimensions as \code{A} such that each
column is sorted increasingly (for \code{sort_each_col}) or contains the
sorting indexes (for \code{sort_index_each_col}).
\code{sort_index_each_col}.
}
\description{
Convenience functions to sort the columns of a matrix in an
increasing way.
}
\keyword{internal}
