% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sobolev.R
\name{sph_stat_Sobolev}
\alias{sph_stat_Sobolev}
\alias{cir_stat_Sobolev}
\title{TODO}
\usage{
sph_stat_Sobolev(X, Psi_in_X = FALSE, p = 0, w = c(0, 0, 1), b = 0)

cir_stat_Sobolev(Theta, Psi_in_Theta = FALSE, w = c(0, 0, 1), b = 0)
}
\arguments{
\item{X}{an \bold{array} of size \code{c(n, p, M)} containing the Cartesian
coordinates of \code{M} samples of size \code{n} of directions on
\eqn{S^{p-1}}. Must not contain \code{NA}'s.}

\item{Psi_in_X}{does \code{X} contain the shortest angles matrix
\eqn{\boldsymbol\Psi}{\Psi} that is obtained with \code{\link{Psi_mat}(X)}?
If \code{FALSE} (default), \eqn{\boldsymbol\Psi}{\Psi} is computed
internally.}

\item{p}{integer giving the dimension of the ambient space \eqn{R^p} that
contains \eqn{S^{p-1}}.}

\item{w}{weights for the (finite) Sobolev test. A non-negative vector.
Defaults to \code{c(0, 0, 1)}.}

\item{b}{bias for the (finite) Sobolev test. Defaults to \code{0}.}

\item{Theta}{a \bold{matrix} of size \code{c(n, M)} with \code{M} samples
of size \code{n} of circular data on \eqn{[0, 2\pi)}. Must not contain
\code{NA}'s.}

\item{Psi_in_Theta}{does \code{Theta} contain the shortest angles matrix
\eqn{\boldsymbol\Psi}{\Psi} that is obtained with
\cr\code{\link{Psi_mat}(array(Theta, dim = c(n, 1, M)))}? If \code{FALSE}
(default), \eqn{\boldsymbol\Psi}{\Psi} is computed internally.}
}
\description{
TODO
}
