% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{utils}
\alias{utils}
\alias{n_from_dist_vector}
\alias{t_inv_sqrt_one}
\title{Low-level utilities for \pkg{sphunif}}
\usage{
n_from_dist_vector(n_dist)

t_inv_sqrt_one(t)
}
\arguments{
\item{n_dist}{a positive integer \eqn{(n - 1) * n / 2} for which \eqn{n}
is to be recovered.}

\item{t}{a vector to evaluate \eqn{t / \sqrt{1 - t^2}}.}
}
\description{
Internal and undocumented low-level utilities for
\pkg{sphunif}.
}
\examples{
n <- 10
sphunif:::n_from_dist_vector((n - 1) * n / 2)
sphunif:::n_from_dist_vector((n - 1) * n / 2 + 1) # Invalid input
}
\keyword{internal}
